refs #145: adds option to XML Schema for HDF5 layout files
Note that this is backwards-compatible in that the default value for
this is "true", matching existing behavior

When present and with value "false", then the HDF5 file plugin is
instructed not to write the default attributes (NDAttrName,
NDAttrSource, NDAttrType) for the designated dataset.

Use this attribute for each dataset that should not have the default
attributes.